+++ AK/NonnullRefPtrVector.h
@@ -31,7 +31,7 @@
 
 namespace AK {
 
-template<typename T, int inline_capacity>
+template<typename T, size_t inline_capacity>
 class NonnullRefPtrVector : public NonnullPtrVector<NonnullRefPtr<T>, inline_capacity> {
 };
